+++ R/diff-time.R
@@ -75,10 +75,7 @@ diff_time <- function(
     wyears = z / getOption("mark.weeks_in_year", 52) / 604800
   )
 
-  class(out) <- c("numeric", "diff_time")
-  attributes(out) <- list(units = method)
-
-  out
+  struct(out, c("numeric", "diff_time"), units = method)
 }
 
 extract_numeric_time <- function(x, tz) {