+++ tests/testthat/test-fact.R
@@ -9,21 +9,21 @@ test_that("fact.pseudo_id() works", {
 test_that("fact() correctly labels NAs [#24]", {
   expect_equal(
     fact(c(NA, "a", "b")),
-    structure(c(3L, 1L, 2L), class = "factor", levels = c("a", "b", NA))
+    struct(c(3L, 1L, 2L), class = "factor", levels = c("a", "b", NA))
   )
 
   expect_equal(
     fact(c(NA, 1, 3)),
-    structure(c(3, 1, 2), class = "factor", levels = c("1", "3", NA))
+    struct(c(3L, 1L, 2L), class = "factor", levels = c("1", "3", NA))
   )
 
   expect_equal(
     fact(c(1, NA, NA, 3)),
-    structure(c(1, 3, 3, 2), class = "factor", levels = c("1", "3", NA))
+    struct(c(1L, 3L, 3L, 2L), class = "factor", levels = c("1", "3", NA))
   )
 
   expect_equal(
     fact(c(TRUE, TRUE, NA, FALSE, TRUE)),
-    structure(c(2, 2, 3, 1, 2), class = "factor", levels = c("TRUE", "FALSE", NA))
+    struct(c(2L, 2L, 3L, 1L, 2L), class = "factor", levels = c("TRUE", "FALSE", NA))
   )
 })