+++ tests/testthat/test-pseudo_id.R
@@ -4,7 +4,7 @@ test_that("pseudo_id.default() works", {
 
   expect_identical(
     res,
-    structure(
+    struct(
       c(1L, 1L, 2L, 4L, 3L),
       uniques = c(1L, 4L, 2L, NA_integer_),
       class = c("pseudo_id", "integer"))
@@ -15,7 +15,7 @@ test_that("pseudo_id.default() works", {
 
   expect_identical(
     res,
-    structure(
+    struct(
       c(1L, 1L, 2L, 4L, 3L),
       uniques = c(1.2, 4.999, -2, NA_real_),
       class = c("pseudo_id", "integer"))
@@ -23,23 +23,23 @@ test_that("pseudo_id.default() works", {
 })
 
 test_that("pseudo_id.factor() works", {
-  x <- structure(1:3, levels = letters[1:3], class = "factor")
+  x <- struct(1:3, levels = letters[1:3], class = "factor")
   res <- pseudo_id(x)
-  exp <- structure(1:3, uniques = letters[1:3], class = c("pseudo_id", "integer"))
+  exp <- struct(1:3, uniques = letters[1:3], class = c("pseudo_id", "integer"))
   expect_identical(res, exp)
 })
 
 test_that("pseudo_id.factor() does not return NA values [#28]", {
   # NA in levels and values
-  x <- structure(c(1:3, NA), levels = c(letters[1:3], NA), class = "factor")
+  x <- struct(c(1:3, NA), levels = c(letters[1:3], NA), class = "factor")
   res <- pseudo_id(x)
-  exp <- structure(1:4, uniques = c(letters[1:3], NA_character_), class = c("pseudo_id", "integer"))
+  exp <- struct(1:4, uniques = c(letters[1:3], NA_character_), class = c("pseudo_id", "integer"))
   expect_identical(res, exp)
 
   # NA in values but not in levels
-  x <- structure(c(1:3, NA), levels = letters[1:3], class = "factor")
+  x <- struct(c(1:3, NA), levels = letters[1:3], class = "factor")
   res <- pseudo_id(x)
-  exp <- structure(1:4, uniques = c(letters[1:3], NA_character_), class = c("pseudo_id", "integer"))
+  exp <- struct(1:4, uniques = c(letters[1:3], NA_character_), class = c("pseudo_id", "integer"))
   expect_identical(res, exp)
 })